+++ Plugins/ProtectedStore/ProtectedStoreException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Excalibur.MvvmCross.Plugin.ProtectedStore
+{
+    public class ProtectedStoreException : Exception
+    {
+
+    }
+}